Rename `cookieDomain` to `persistDeviceBetweenDomain`, also created some helper functions for cookie
Also moved all deviceIdentification code to `getDeviceIdentification()` so we have 1 place where we can get it from
Also removed subscribing device upon constructing App.jsx, because it doesn't make sense to register a device when the user does not have interacted with the chat yet. That's why we subscribe after we show the chat.