Fix issue where WorkingHours.js would only check for the current day and not day+month+year
We checked if the current timestamp is greater than the end timestamp AND the day numbers are the same.
But we should've checked if the DATES are the same, because if they are that means we are in the same day/date as the timestamp and if that is the case that means we are OUTSIDE the office hours (the office hours only count for that day).
Also added tests for this