Keep track of the existing registration
Previously we did this using the localStorage, but removed that. Now we keep track in the current proces.

When you re-open the chat by clicking on it, we don't register a new device (worked previously like this, but now it doesn't use the localStorage anymore to track the registration).

When you refresh the page we lose the memory and have to re-register again. Previously we used the information stored in the localStorage to re-register, but we don't want that anymore. The only thing we use is the `deviceIdentification` so we don't generate a new one on each registration.