Limit local storage information (#38)
* removed unnecessary data from localstorage creation

* Added cypress test to check if localstorage data still exist or not

* Updated package-lock json

* Updated package-lock json

* Fixed missing dependencies from package-lock json

* Removed unnecessary function for test

* Updated package-lock.json

* split test in two seperated tests

* updated code based on feedback and eslint

* Remove check for existing device information when registering

When we register a device we always want to use the uptodate information. Also this check wasn't needed because the register function is only called when something, related to the device information, changed (in `shouldComponentUpdate`).

* Remove unnecessary clearing of localStorage before registering

We don't need to do this anymore, because we don't check the local storage when we register.

* Keep track of the existing registration

Previously we did this using the localStorage, but removed that. Now we keep track in the current proces.

When you re-open the chat by clicking on it, we don't register a new device (worked previously like this, but now it doesn't use the localStorage anymore to track the registration).

When you refresh the page we lose the memory and have to re-register again. Previously we used the information stored in the localStorage to re-register, but we don't want that anymore. The only thing we use is the `deviceIdentification` so we don't generate a new one on each registration.

* Remove colors from cypress output

---------

Co-authored-by: boukeTracebuzz <bouke.de.vries@tracebuzz.com>