Move messengerOpenState to state and save any changes to it in localStorage
This way the localStorage is always up-to-date, and we can now use the value from the state instead of inferring it from `this.state.showChat`.

Also renamed `messengerState` to `messengerOpenState` because it might insinuate that it is the `state` of the component, which is not true..

Added some debug logging for this, so we can easily debug production deployments

Also fixed a test that didn't wait for the chat to be initialized.