Hide chat instead of unmounting
This way we can keep the "state" between hiding and showing the chat. This is especially helpful with the new feature where we wait for a subsribe to complete before sending your message. You can submit your message during the subscribe call, but it will be sent after the subscribe call is done. If you were to close the chat (minimize) and reopen, you would have an empty input field and was able to submit another message.

So this fixed it by only hiding the chat (`display: none`) so we can keep the state.