Updated test for apiCustomHeaders
a bit out of scope for this branch, but i didn't want to make a whole new ticket for this. The custom headers are always merged with the previous values, but we didn't test this specifically. We only tested that the new headers were set, so i added a test which does this.