Wait for device registration when submitting message (#45)
* Hide chat instead of unmounting

This way we can keep the "state" between hiding and showing the chat. This is especially helpful with the new feature where we wait for a subsribe to complete before sending your message. You can submit your message during the subscribe call, but it will be sent after the subscribe call is done. If you were to close the chat (minimize) and reopen, you would have an empty input field and was able to submit another message.

So this fixed it by only hiding the chat (`display: none`) so we can keep the state.

* Fix issues with icon tests

* Better way of waiting for the api call to finish

* Updated test for apiCustomHeaders

a bit out of scope for this branch, but i didn't want to make a whole new ticket for this. The custom headers are always merged with the previous values, but we didn't test this specifically. We only tested that the new headers were set, so i added a test which does this.

* Re-register, on chat open, if previous register failed

If you register and that fails, we want to be able to do that call again upon re-opening the chat.