Fix console error that Launcher messengerOpenState cannot be null
This was caused due to the initial value of `messengerOpenState` being `null`. I have set it to an openState that is decided on by looking at `showChat`.

We also need to save this initial value into the localStorage, because the `shouldComponentUpdate` won't do that. This is because the default value is set in the `constructor` which doesn't trigger a state update. So i added the saving into localStorage into `componentDidMount` as that is ran once after initializing.