Fix merging issue with weekdays
The problem is that `deepMerge` was adding the 2 arrays together. When you executed `window.parleySettings.weekdays = [["Monday", 0.00, 1.00]]` twice, you ended up with this in your state `[["Monday", 0.00, 1.00], ["Monday", 0.00, 1.00]]`. To fix this is added some code that looks at the value for the first index for each entry. If this value already exists in the state we remove it. This way you cannot have duplicates. This also works for the `[timestamp, timestamp]` format.