Add Reply button functionality and tests
I also needed the ref for ReplyActions so i could set the `reply` in the state. For this i needed to remove the `forwardRef()`. After removing this all the tests still worked and also the ref (for which it was added i guess) still worked. So i think this wasn't necessary anymore...