Add rich message buttons (#60)
* Add support for rendering buttons

* Add onclick support for CallButton and WebUrlButton

* Add test for WebUrl button

* Add test for Call button

* Add Reply button functionality and tests

I also needed the ref for ReplyActions so i could set the `reply` in the state. For this i needed to remove the `forwardRef()`. After removing this all the tests still worked and also the ref (for which it was added i guess) still worked. So i think this wasn't necessary anymore...

* Remove .only

* Make titles required

* Send the message onClick instead of only setting the text

* Show payload as button title when there is no title

* Remove todo

* Fix issue with empty arrow functions

* Let the reply button call the api directly

There is no need to try and reuse the "submitHandler" from ResultActions, because the reply button doesn't need to
- disable the text field
- check if the payload is empty (it is never empty thanks to the validation done in the api)
- put focus on the text area (that wasn't the last place the client focused on)

* Fix image location in message

* Move title logic to button

* Let buttons open using "noopener"

This is to prevent https://owasp.org/www-project-web-security-testing-guide/latest/4-Web_Application_Security_Testing/11-Client-side_Testing/14-Testing_for_Reverse_Tabnabbing

* Remove .only

* Fix long line

* Move button styling to its own css module

* Add `noreferrer` to window.open

This is to prevent the following https://owasp.org/www-project-web-security-testing-guide/latest/4-Web_Application_Security_Testing/11-Client-side_Testing/14-Testing_for_Reverse_Tabnabbing