Load interfaceTexts overrides at the very end (#63)
This way you can always override any text. This was not possible for example `title` because it was set after the override and if you didn't set `desc` it would be filled with `title`'s default.

Also added tests for these 'new names' and made the old names (that were missing) reactive.