Add key to Chat so that it re-renders when account changes
Previously we always did a full GET /messages call with all message in it. When account identification changes so did al the messages. We got `[]` back and rendered that. But now we almost always get `[]` because we use the GET /messages/after:xxx so this didn't work anymore.

Not the cleanest but the quickest solution.