Use new client-api endpoint for counting unread messages  (#69)
* update browserlist

* update client api version number

* Basics of new message counting

Added GET /messages/unseen/count api call.
Added PUT /messages/status/:newStatus api call.
Using these new calls in the UI instead of locally tracking unread status
Added an extend on Polling which is used for unread messages count polling
Have not made any tests yet. Also have not tested everything yet in UI..

* add todos

* Update fixtures

* Dont abuse restartPolling for startPolling

* Move marking messages as read to where they are actually rendered

* All tests should work now

* Bigint is causing some weird issues

and honestly i don't mind if we don't have a test for this type...

* Big changes to polling

Made sure we always have 1 polling service active. There are 3 services:
- message polling (GET /messages)
- unread message polling (GET /messages/unread/count)
- unread message slow polling (same as unread but slow, meant for when the chat has never been opened in this session)

Switch to the faster unread messages polling if the chat goes from open to closed

Removed handling of new messages from App.jsx because it is not needed anymore since we can fully rely on the unread count to see if there are new messages

* Oops, this shouldn't have been committed

* Fix test

* Fix test

This one was caused by the fact that we updated the `this.Api` with new params but we didn't start the new MessagePollingService. We started an old copy of it.

* Add unit tests

* Remove todo

* Fix issue in polling test