Render welcome-message more intelligent (#70)
* Render welcome-message more intelligent

Render underneath the date if it is the same as the current day

Else; render it at the bottom (above the sticky message).

* Remove duplicate check

* Fix flaky tests

The /messages/before:123 was not captured so it gave back a different Welcome message than what we expect. This could happen right when we compare the welcome message text and thus fail the test. Sometimes the check was before/after this other welcome message and it succeeded