Fix issue "Scrollable region must have keyboard access"
The div with role `feed` in Conversation.jsx did not have a valid tab index. Not sure why we needed the -1, but i replaced it with a correct index. I also reordered the other indexes so the order is better. It is now:
1. Input field
2. Conversation
3. Minimize button
4. Launcher button

See https://dequeuniversity.com/rules/axe/4.10/scrollable-region-focusable?application=axeAPI