Fix issue "Certain ARIA roles must contain particular children"
The conversation node with role=feed must have articles in them. When the conversation has not yet been started we need to fill it with something. This fills the conversation with an article saying that there are no messages yet. I hide it so it is only used for screen readers. I could've made this visible to all users, but that would need some more testing and better design which is out of scope