Fixed axe issue "Ensure elements with an ARIA role that require child roles contain them"
The problem was that children in a role=feed node can only be role=article. So i made the carousel role=article, but then we got the same error for the navigation buttons. They are not allowed in a role=feed directly, so we make the container also a role=article.