Improve chat accessibility (#71)
* Add cypress-axe

* Fix tests so they don't crash cypress-axe

* Add inspection for missing accessibility test

* Fixed tests

* Add test source

* Add new setting 'alwaysShowSendButton'

Enabling this setting will always show the send button even when you are not on a mobile device. This is required to pass https://www.w3.org/WAI/WCAG21/Understanding/labels-or-instructions

* Add attributes to chat window

These are needed for the WCAG21 Success Criterion 4.1.2: Name, Role, Value

Using techniques:
- https://www.w3.org/WAI/WCAG21/Techniques/aria/ARIA4
- https://www.w3.org/WAI/WCAG21/Techniques/aria/ARIA5
- https://www.w3.org/WAI/WCAG21/Techniques/aria/ARIA14

* Announce new messages

* Add translations for screen reader announcement

* Make sure announcer is created

* Make chat window accessible on 500% zoom

500% zoom is the max on Firefox and Chrome. When using this (or something like 400%) the chat was unusable. Now we activate the "mobile" styling when we pass a certain threshold. We don't activate other mobile features since they are not needed in this case.

* Ignore env file

* Replace cypress-axe with @axe-watcher

This works way better and doesn't inject stuff which might violate the CSP.

For this we needed to update to at least node v16.

* Fix issue "Elements must only use permitted ARIA attributes"

The button container for message had an aria-label but no role attribute, which is required in some cases. See https://dequeuniversity.com/rules/axe/4.10/aria-prohibited-attr?application=axeAPI

* Make sure axe only checks the app and not the whole page

* Fix issue "Scrollable region must have keyboard access"

The div with role `feed` in Conversation.jsx did not have a valid tab index. Not sure why we needed the -1, but i replaced it with a correct index. I also reordered the other indexes so the order is better. It is now:
1. Input field
2. Conversation
3. Minimize button
4. Launcher button

See https://dequeuniversity.com/rules/axe/4.10/scrollable-region-focusable?application=axeAPI

* Fix issue "Certain ARIA roles must contain particular children"

The conversation node with role=feed must have articles in them. When the conversation has not yet been started we need to fill it with something. This fills the conversation with an article saying that there are no messages yet. I hide it so it is only used for screen readers. I could've made this visible to all users, but that would need some more testing and better design which is out of scope

* Fix issue "Buttons must have discernible text"

Just added an aria-label to the media download button

* Fix issue "Elements must meet minimum color contrast ratio thresholds"

The ratio of the link color compared to the background was not enough. So i changed the color.

* Fix issue "Ensure links are distinguished from surrounding text in a way that does not rely on color"

Added underline to the links as this is the only way to distinguish it from the surrounding text and keep the link color blue.

* Fix issue "Buttons must have discernible text"

Added aria-labels to buttons that were missing

* Fix issue "Scrollable region must have keyboard access"

The carousel was not focusable. I added a tabindex to it and reset all tabindexes to 0 so they get the DOM order instead. This might be easier to understand for users instead of a custom order. Also a custom order would be harder to maintain.

* simplify if

* Split up UI tests in separate files

* Remove .only

* Fix flaky test

I have no idea why this works but it does.
Without it you sometimes get an error on `scrollIntoView`, something about not being able to find the parentElement

* Make sure only one option can render

* Fix some tests

* Ignore viewport in axe test

* Fixed axe issue "Ensure elements with an ARIA role that require child roles contain them"

The problem was that children in a role=feed node can only be role=article. So i made the carousel role=article, but then we got the same error for the navigation buttons. They are not allowed in a role=feed directly, so we make the container also a role=article.

* Remove exclude since it doesn't work

* Show axe violations in cypress

* Fix another "Ensure elements with an ARIA role that require child roles contain them" issue

This time the anchor tag in the welcome message is not allowed directly in a feed. So i wrapped the whole welcome message in an article

* Fix axe isse "Ensure <meta name="viewport"> does not disable text scaling and zooming"

Not sure why we had this `user-scalable=no`, but i removed it. The tests all work fine without it.

* Adapt ci config to @axe-core/watcher

* Use chrome by default

* Ignore @babel/no-invalid-this

This is not interesting to fix in the tests

* Fix stylint issues

* Remove unused comments

* Remove unnecessary plugin code

* Fix wrong defaults in css

* Add missing css vars to parleycdn index