Make upload button accessible (#73)
* Make upload button accessible

The upload button was not selectable using the TAB key. This was due to the fact that we were using a <label> as the button. This however is not selectable with TAB. I could've done this in plain js but that would mean i have to manually trigger the file input when clicking on a button. Using `react-aria-components` they do all of that for us.